delete posts coresponding to user

+++ app/api/bikes/[id]/route.ts
@@ -1,6 +1,8 @@
 import prisma from "../../../../prisma/index";
 import { NextResponse } from "next/server";
 import { Prisma } from "@prisma/client";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 type ParamsType = {
   params: {
@@ -35,8 +37,27 @@ export async function GET(request: Request, { params }: ParamsType) {
 }
 
 export async function DELETE(request: Request, { params }: ParamsType) {
+  const session = await getServerSession(authOptions);
+
+  if (!session) {
+    return new NextResponse("Login to delete a bike", { status: 401 });
+  }
+
   try {
     const { id } = params;
+
+    const findBike = await prisma.bike.findUnique({
+      where: {
+        id: id,
+      },
+    });
+
+    if (findBike?.userId !== session.user.id) {
+      return new NextResponse(
+        "This is not your bike. You can't delete this bike.",
+        { status: 401 }
+      );
+    }
 
     const data = await prisma.bike.delete({
       where: {
@@ -47,11 +68,9 @@ export async function DELETE(request: Request, { params }: ParamsType) {
     if (!data) {
       throw new Error();
     }
+
     return NextResponse.json(data);
   } catch (error: any) {
-    if (error instanceof Prisma.PrismaClientKnownRequestError) {
-      console.log("aici", error);
-    }
     return new NextResponse("Error has occured while deleting this bike", {
       status: 403,
     });
